Removes several files from the repo to clean up git-status
These files are altered by the game automatically, due to writing timestamps or other things, and do not need to be tracked, as such, they were removed and added to .gitignore
These will be generated by the mods if missing.
I believe that only IC2.ini will show up as modified in git-status now, as it also saves timestamps.